+++ src/Microsoft.TestPlatform.Extensions.BlameDataCollector/DumpTypeOption.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.TestPlatform.Extensions.BlameDataCollector
+{
+    public enum DumpTypeOption
+    {
+        Full,
+        WithHeap,
+        Mini,
+    }
+}
