+++ src/Microsoft.TestPlatform.Extensions.BlameDataCollector/ICrashDumper.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.TestPlatform.Extensions.BlameDataCollector
+{
+    public interface ICrashDumper
+    {
+        void AttachToTargetProcess(int processId, string outputFile, DumpTypeOption dumpType);
+
+        void WaitForDumpToFinish();
+
+        void DetachFromTargetProcess(int processId);
+    }
+}
