+++ src/Microsoft.TestPlatform.Extensions.BlameDataCollector/IHangDumper.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.TestPlatform.Extensions.BlameDataCollector
+{
+    public interface IHangDumper
+    {
+        void Dump(int processId, string outputFile, DumpTypeOption dumpType);
+    }
+}
